@@ -14,24 +14,26 @@
   "license": "ISC",
   "devDependencies": {
     "@babel/cli": "^7.17.6",
-    "@babel/core": "^7.17.8",
+    "@babel/core": "^7.27.4",
     "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-env": "^7.27.2",
+    "@babel/preset-react": "^7.27.1",
     "@babel/runtime": "^7.17.8",
     "@viewlift/player": "^2.0.14",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.4",
+    "babel-loader": "^8.4.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "css-loader": "^7.1.2",
     "dev-player": "file:../packages/player-core/src",
-    "webpack": "^5.71.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.0"
+    "html-webpack-plugin": "^5.6.3",
+    "style-loader": "^4.0.0",
+    "webpack": "^5.99.9",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.15.2"
   },
   "dependencies": {
-    "css-loader": "^6.7.3",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "style-loader": "^3.3.1",
     "videojs-contrib-quality-menu": "^1.0.4",
     "videojs-dvr": "^1.4.0"
   }
